Update tutorial to clarify that blog be a child of the homepage
`Under the homepage` [implies to some](https://github.com/wagtail/wagtail/issues/1583)  that the Blog be at the same level as the home page, listed below (under) it (see #1583). If you do this, you will not be able to see your `/blog` url, and the error warning you receive (to configure a site), will not lead to anywhere you can assign a site to the blog.